refactor(auto_search): extract pack_detection.rs — Phase 1 complete
Fifth and final slice of the Phase 1 auto_search.rs split. Extracted the
sibling-series detection + transitive-relation walk cluster:

  pub struct SiblingMatch
  pub(super) fn is_pack_candidate_relation
  pub const TRANSITIVE_WALK_MAX_FETCHES
  pub fn is_transitive_walk_source
  pub fn transitive_relation_graft
  pub fn expand_parent_with_transitive_relations
  pub fn detect_sibling_entries_in_pack      (+ helpers)
  fn compute_sibling_episode_offset
  fn detect_sibling_via_episode_range

  + 26 #[test] fns covering the detector, fallback paths, transitive
    graft, and the `related` / `detail_with_titles` helpers they need
    (duplicated in both release_parse.rs and pack_detection.rs test
    modules since both subsystems exercise the same AnimeDetail shape).

Module wiring:
  - pack_detection.rs imports from super: is_media_filename,
    normalize_subtitle, trailing_subtitle_of, within_episode_slack
    (all pub(super) in release_parse, re-exported through mod.rs).
  - aliases.rs now reaches is_pack_candidate_relation via explicit
    `use super::pack_detection::is_pack_candidate_relation;` rather
    than a re-export — the function is pack-detection-internal and
    aliases is its only external caller.
  - mod.rs pub use surface for external callers preserved:
    TRANSITIVE_WALK_MAX_FETCHES, detect_sibling_entries_in_pack,
    expand_parent_with_transitive_relations, is_transitive_walk_source
    (all consumed by handlers / auto_expand / main).

── Phase 1 final tally ────────────────────────────────────────────────

Before: src/services/auto_search.rs  — 6557 lines, one file.
After:  src/services/auto_search/
          mod.rs             — 2278 lines
          pack_detection.rs  — 2089 lines
          release_parse.rs   — 1064 lines
          aliases.rs         —  707 lines
          search_target.rs   —  321 lines
          scoring.rs         —  287 lines
          ─────────────────────────────
          total              — 6746 lines (190-line growth from
                                 module headers, pub use blocks,
                                 duplicated test helpers)

No file over 2300 lines. External call sites (main.rs, rss.rs,
auto_expand.rs, handlers/library/search.rs) unchanged — every
auto_search::X path still resolves via pub use re-exports in mod.rs.
Tests: 601 passing, 0 failing.
